requirements.txt
requests==2.31.0
feedparser==6.0.11
beautifulsoup4==4.12.2
Jinja2==3.1.2
pydantic==2.5.3
python-dateutil==2.8.2
pytz==2023.3
